Using google aio storage and aiohttp
