+++ criterion_core/utils/gcs_downloader.py
@@ -1,13 +1,8 @@
-import asyncio
-import aiohttp
-from gcloud.aio.storage import Storage
-from tensorflow.python.lib.io import file_io
 import os
 import cv2
 import numpy as np
 import time
 import multiprocessing
-import queue
 import random
 
 
@@ -73,85 +68,6 @@ def random_affine_transform(images,
         return im_t
 
     return [transform(im) for im in images], A
-
-
-def gcs_walk(bucket_name, content_filter: str = "image",
-             service_file: str = os.environ.get('GOOGLE_APPLICATION_CREDENTIALS', r'urlsigner.json')):
-    async def async_walk():
-        conn = aiohttp.TCPConnector(limit_per_host=30)
-        async with aiohttp.ClientSession(connector=conn) as session:
-            st = Storage(service_file=service_file, session=session)
-            blobs = await st.list_objects(bucket_name)
-            output = [blob for blob in blobs['items'] if content_filter in blob["contentType"]]
-        return output
-    loop = asyncio.get_event_loop()
-    out = list(loop.run_until_complete(asyncio.wait((async_walk(), )))[0])[0].result()
-    return out
-
-
-class GcsDownloader(multiprocessing.Process):
-    def __init__(self, q_in, q_done, service_file="urlsigner.json", **kwargs):
-        multiprocessing.Process.__init__(self)
-        if not os.path.exists(service_file):
-            file_io.copy("gs://security.criterion.ai/urlsigner.json", service_file)
-        self.service_file = service_file
-        self.q_in = q_in
-        self.q_done = q_done
-        self.loop = None
-        self.kwargs = kwargs
-        self.start()
-
-    @staticmethod
-    async def download_file(bucket_name, blob, st):
-        for retries in range(10):
-            try:
-                byte_buffer = await st.download(bucket_name, blob, timeout=100)
-                return byte_buffer
-            except aiohttp.ClientResponseError:
-                pass
-        raise aiohttp.ClientResponseError
-
-    async def download_loop(self):
-        conn = aiohttp.TCPConnector(limit_per_host=30)
-        async with aiohttp.ClientSession(connector=conn) as session:
-            st = Storage(service_file=self.service_file, session=session)
-            while True:
-                blobs = self.q_in.get()
-                futures = [GcsDownloader.download_file(blob['bucket'], blob['name'], st) for blob in blobs]
-                buffers = await asyncio.gather(*futures)
-                self.q_done.put((buffers, [blob['class'] for blob in blobs]))
-                self.q_in.task_done()
-
-    @property
-    def loop(self):
-        return self._loop
-
-    @loop.setter
-    def loop(self, value):
-        self._loop = value
-
-    @loop.getter
-    def loop(self):
-        if self._loop is None:
-            self._loop = asyncio.new_event_loop()
-            asyncio.set_event_loop(self._loop)
-        return self._loop
-
-    def reset_queues(self):
-        while True:
-            try:
-                self.q_in.get_nowait()
-            except queue.Empty:
-                break
-        while True:
-            try:
-                self.q_done.get_nowait()
-            except queue.Empty:
-                break
-
-    def run(self):
-        self.loop.run_until_complete(asyncio.wait((self.download_loop(),)))
-
 
 
 def main(bucket_name="4679259e-7a0a-4e85-90cf-a52f3451cf38.datasets.criterion.ai",
